Weird refresh in list of smses showing only numbers for a second
Bug 5649763
When deleting a thread, the code was rebuilding the whole contacts cache
and threads would only show numbers while the contacts were loading. This
is normally only visible for folks with a LOT of conversations. With this
change, deleting a single thread only removes the associated contact(s)
from the cache.

Change-Id: Ia31f9fdcb3654461d93ed5de205a28ab5389a865